Extract internal search refs classes
