+++ src/us/kbase/workspace/database/refsearch/ReferenceSearchMaximumSizeExceededException.java
@@ -1,0 +1,16 @@
+package us.kbase.workspace.database.refsearch;
+
+/** An exception thrown when a search has traversed the maximum number of references allowed.
+ * @author gaprice@lbl.gov
+ *
+ */
+@SuppressWarnings("serial")
+public class ReferenceSearchMaximumSizeExceededException extends Exception {
+	
+	/** Construct the exception.
+	 * @param message an exception message.
+	 */
+	public ReferenceSearchMaximumSizeExceededException(final String message) {
+		super(message);
+	}
+}
